Update working-directory in Github Actions workflow
Updated the `working-directory` for the test run command in the Github Actions workflow. It was previously hard-coded to `src/UnitTestProject`, but now it dynamically uses the `projectName` environment variable. The new directory is `src/${{ env.projectName }}.UnitTests`. This change provides more flexibility for changing project names or handling multiple projects.