Update Github Actions workflow and .NET version
This commit includes several updates to the Github Actions workflow and .NET version. The `runs-on` parameter for the `build` job has been changed to `windows-latest` and a new `test` step has been added to this job. This step runs `dotnet test` with specific parameters for logging and code coverage. An additional step has been added to upload the build artifact using the `actions/upload-artifact@v4` action. The `runs-on` parameter for the `test` job has been changed to `ubuntu-latest` and the `test` step in this job has been removed. Lastly, the `dotnet-version` for the `Setup .NET` step has been specified as `8.0.x`.