Update Github Actions workflow and artifact handling
Updated the Github Actions workflow configuration, specifically for the jobs related to build artifacts and test results. The artifact name for upload and download has been changed from 'drop' to 'test-results', with the path set to the TestOutput.xml file. A new step has been added to upload a 'test-coverage' artifact, with the path set to the coverage.cobertura.xml file. The step to publish test results is now configured to always execute, using the action 'EnricoMi/publish-unit-test-result-action@v2'.