Update artifact path in Github Actions workflow
The Github Actions workflow file has been updated to modify the path of the build artifact in the "Upload a Build Artifact" job. The path has been changed from `outputs.coverageFile` to `needs.build.outputs.coverageFile`, indicating that the `build` job is now responsible for generating the coverage file. The location of this file is then passed to the "Upload a Build Artifact" job.