Update PowerShell script in GithubActions.yml for coverage file path
The GithubActions.yml file has been updated to improve the handling of the coverage file path. The PowerShell script now uses the `Write-Host (Get-Location)` command to print the current location. The command to find the coverage file has been modified to store the path of the most recently modified 'coverage.cobertura.xml' file in the 'TestResults' directory in the `$coverageFile` variable. This path is then written to the GitHub output environment variable using `Write-Output "coverageFile=$coverageFile" >> $env:GITHUB_OUTPUT` and printed using `Write-Host $coverageFile`. The previous command to echo the path of the coverage file to the GitHub output environment variable has been commented out.