`Add step to find coverage file in GitHub Actions workflow`
This commit introduces a new step in the GitHub Actions workflow, specifically in the GithubActions.yml file. The step, identified as `find-coverage-file`, is designed to locate the most recent 'coverage.cobertura.xml' file in the './TestResults' directory of the project's unit tests. This file is expected to contain the code coverage results from the unit tests. The step uses PowerShell as the shell and the working directory is set to the unit tests directory of the project.