Update XML file path in `$xmlSource` variable
Updated the path to the XML file in the `$xmlSource` variable. It now directly references the `coverage.cobertura.xml` file in the current directory, instead of using the `artifactStagingDirectory` environment variable. This change was likely made to accommodate the XML file being generated or located in the same directory as the running script.