Update artifact name in Github Actions workflow
Updated the name of the artifact in the "Upload a Build Artifact" step of the Github Actions workflow. The artifact name has been changed from `test-coverage` to `test-coverage-summary`. This change only affects the name of the uploaded artifact after the job is run, with the path for the artifact remaining the same. No other changes have been made to the workflow.