Refactor Github Actions workflow for test coverage
This commit includes several changes to the Github Actions workflow file. The step to download build artifacts has been moved from the `jobs` section to the `publish-test-coverage` section, meaning that artifacts are now downloaded only when the `publish-test-coverage` job is run. The `if: always()` condition has been removed from the `jobs` section, and the `if: success()` condition has been added to the `publish-test-coverage` section, ensuring that this job only runs if previous jobs were successful.

Additional parameters have been added to the `Download Build Artifacts` step in the `publish-test-coverage` section, specifying that the `test-coverage` artifact will be downloaded to the `artifactStagingDirectory` environment variable path. Lastly, a new step named `Code Coverage Summary` has been added to the `publish-test-coverage` section, which runs a PowerShell script to print the current location for debugging or informational purposes.