`Remove 'Add Markdown to summary' job and modify 'Collect job summary' job`
Removed the `Add Markdown to summary` job that was responsible for reading and appending the content of `coverage.cobertura.md` to the `GITHUB_OUTPUT` environment variable. Modified the `Collect job summary` job to directly append the string "Foo Bar" to the `GITHUB_STEP_SUMMARY` environment variable, instead of echoing the output of the removed job.