Uncommented Publish Test Results job and added new command in Collect job summary
The `Publish Test Results` job in the GithubActions.yml file has been uncommented. This job uses the `EnricoMi/publish-unit-test-result-action@v2` action and is always executed, publishing the test results found in the `TestOutput.xml` file. Additionally, a new command has been added to the `Collect job summary` job. This command uses `Get-Content` to read and echo the content of the `coverage.cobertura.md` file, potentially useful for debugging or logging.