Disable test results publishing in Github Actions
In this commit, the "Publish Test Results" job in the GithubActions.yml file has been commented out. This includes the `EnricoMi/publish-unit-test-result-action@v2` action, the `if: always()` condition, and the `with:` clause specifying the test results file path. This change indicates that the publishing of test results is temporarily disabled.