Update string appending in GithubActions.yml
Updated the method of appending the string "Foo Bar Bar" to the `GITHUB_STEP_SUMMARY` environment variable in the GithubActions.yml file. The previous method used the `echo` command to directly append the string "Foo Bar". The new method uses the `Add-Content` cmdlet in PowerShell, providing more control over the process, including the ability to specify the encoding type, which is now set to utf8.