Update job summary processing in GithubActions.yml
Refactored the way job summary information is collected and processed in GithubActions.yml. Instead of echoing the content from "coverage.cobertura.md" line by line, the updated code now stores this content in a variable named '$content' and adds it to the GITHUB_STEP_SUMMARY environment variable. Removed the 'Foo Bar Bar' string from the GITHUB_STEP_SUMMARY, suggesting it was unnecessary or replaced by the "coverage.cobertura.md" content. The encoding for the content remains as utf8.