Update dependencies and add env variable in GithubActions.yml
Updated the `Publish Coverage Results` job in the GithubActions.yml file to include a new dependency on the `new-version` job. An environment variable `buildVersion` has been added to this job, which is set to the output of the `new-version` job. This ensures that the `Publish Coverage Results` job uses the build version generated by the `new-version` job.