Update artifact upload path in Github Actions
Modified the path for the build artifact upload in the Github Actions workflow. The previous path included the build version in the filename (`*.${{ env.buildVersion }}.nupkg`). The updated code removes this specification, and now any `.nupkg` file in the artifact staging directory will be uploaded (`*.nupkg`). This change provides more flexibility in the naming of the build artifacts.