Update artifact upload step in GitHub Actions
Updated the step that uploads the build artifact in GitHub Actions. The name of the step has been changed to "Upload Build Artifacts" and the action used is now "actions/upload-artifact@v4". Parameters for this action have been specified in the "with" section, setting the artifact name to "build-artifact" and the upload path to all .nupkg files in the artifact staging directory.