Updated GithubActions.yml with new job and utf8 encoding
This commit includes significant updates to the GithubActions.yml file. The encoding for adding content to the GitHub Actions job summary has been updated to utf8, ensuring correct content display in the workflow run details. A new job, 'Publish Build Artifacts', has been added which runs on the latest Ubuntu version and is executed if the 'build' and 'new-version' jobs are successful. This job includes steps for checking out the repository, downloading build artifacts from Windows to Ubuntu, and pushing NuGet packages to the NuGet feed.