Update job configuration in GithubActions.yml
Updated the "Publish Build Artifacts" job configuration in GithubActions.yml for publishing build artifacts to the NuGet feed for production release. The job is set to run on the latest Ubuntu version and in the '.NET Libraries' environment. Added a condition for the job to run only if the previous jobs have been successful. The job now depends on 'build' and 'new-version' jobs, meaning these jobs need to complete successfully before the 'Publish Build Artifacts' job can run.