Update job dependencies and add new-release job in GithubActions.yml
The `build` job has been renamed to `new-build` and is now responsible for building, testing, and packaging the project on a Windows platform. The `publish-test-results`, `publish-test-coverage`, and `Publish Build Artifacts` jobs now depend on `new-build` and `new-version` jobs. The step to download build artifacts for test results in `Publish Build Artifacts` job has been renamed to `Download Build Artifacts`. A new job `new-release` has been added, which creates a new GitHub release and tags the new version. It runs on a Windows platform and depends on `new-build` and `new-version` jobs.