Updated job dependencies and release creation in GithubActions.yml
Updated the GithubActions.yml file to remove the environment for the `publish-build-artifacts` job and made `new-build` and `new-version` jobs prerequisites for it. The step to create a GitHub release and tag now includes all files in the `artifactStagingDirectory` instead of just .nupkg files. Also, the `GITHUB_TOKEN` environment variable is now being used for authentication in this step.