Update GithubActions.yml for Ubuntu and NuGet files
Updated the `runs-on` parameter in the `new-release` job of the GithubActions.yml file to run on the latest version of Ubuntu instead of Windows. Also, modified the `files` parameter in the `Create GitHub Release & Tag` step to include only `.nupkg` files from the `artifactStagingDirectory`, meaning only NuGet package files will be included in the GitHub release and tag.