Update artifact handling and remove release creation
This commit includes significant changes to the handling of build artifacts and the creation of GitHub releases. The `actions/download-artifact@v4` action is now used to download build artifacts to the path specified in the `artifactStagingDirectory` environment variable. The step that listed the contents of the `artifactStagingDirectory` using the `dir` command has been removed. Additionally, the step to create a GitHub release and tag using the `softprops/action-gh-release@v2` action has been removed. This step previously created a GitHub release and tag with the version specified in the `buildVersion` environment variable, using `.nupkg` files from the `artifactStagingDirectory`.