Added new ItemGroup to CommandBridge.csproj and removed CommandBridge.sln
A new ItemGroup was added to the CommandBridge.csproj file to include an image file (icon.png) in the package. The image is located at ../../images/icon.png and is set to be included in the package with the path content\icon.png. Additionally, the CommandBridge.sln file was completely removed, indicating that the solution is no longer needed or is being replaced.