Added new project ".images" and solution item
Added a new project named ".images" to the solution file "CommandBridge.sln" with the associated GUID "{13C58B25-7E88-4CEC-B050-D905F5475A97}". Also, a new item has been added to the "SolutionItems" section of the ".images" project, which is a reference to the file "..\images\icon.png". No code removals were made in this commit.